Get 64-bit build working

+++ xll/utility/string.h
@@ -3,11 +3,11 @@
 #include <string>
 #include "Stringapiset.h"
 
-inline std::basic_string<wchar_t> MultiByteToWideChar(const char* s, size_t n = 0)
+inline std::basic_string<wchar_t> MultiByteToWideChar(const char* s, int n = 0)
 {
 
 	if (n == 0)
-		n = ::strlen(s);
+		n = static_cast<int>(::strlen(s));
 		
 	int n_ = MultiByteToWideChar(CP_ACP, 0, s, n, 0, 0);
 	std::basic_string<wchar_t> s_(n, 0);
@@ -16,10 +16,10 @@ inline std::basic_string<wchar_t> MultiByteToWideChar(const char* s, size_t n = 
 	return s_;
 }
 
-inline std::basic_string<char> WideCharToMultiByte(const wchar_t* s, size_t n = 0)
+inline std::basic_string<char> WideCharToMultiByte(const wchar_t* s, int n = 0)
 {
 	if (n == 0)
-		n = ::wcslen(s);
+		n = static_cast<int>(::wcslen(s));
 		
 	int n_ = WideCharToMultiByte(CP_ACP, 0, s, n, 0, 0, 0, 0);
 	std::string s_(n_, 0);
